Added GrayScale Operation and Code Cleanup
Added GrayScale Operation and Code Cleanup

+++ Pixel/Pixel/Halftoning.cpp
@@ -71,7 +71,7 @@ namespace Pixel
 						{
 							xx = 0;
 						}
-						else if(xx >= width)
+						else if(xx >= (int)width)
 						{
 							xx = width - 1;
 						}
@@ -80,7 +80,7 @@ namespace Pixel
 						{
 							yy = 0;
 						}
-						else if(yy >= height)
+						else if(yy >= (int)height)
 						{
 							yy = height - 1;
 						}
